lets go

+++ 0x05-python-exceptions/1-safe_print_integer.py
@@ -1,7 +1,7 @@
 #!/usr/bin/python3
 def safe_print_integer(value):
     try:
-        print("{:d}".format(float(value)))
+        print("{:d}".format(int(value)))
         return True
     except (TypeError, ValueError) as e :
         return False
